VijaraSocialTek/vajara_ai/requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cycler==0.11.0
Deprecated==1.2.13
Django==4.0.4
et-xmlfile==1.1.0
excel2json==1.0.1
excel2json-3==0.1.6
firebase-admin==5.2.0
fonttools==4.33.3
gcloud==0.18.3
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.1
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
Js2Py==0.71
jwcrypto==1.3
kiwisolver==1.4.2
matplotlib==3.5.2
msgpack==1.0.3
numpy==1.22.3
oauth2client==4.1.3
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pdfkit==1.0.0
Pillow==9.1.1
proto-plus==1.20.4
protobuf==4.21.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pyjsparser==2.7.1
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==3.3.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
wget==3.2
wrapt==1.14.1
xlrd==2.0.1
